requirements.txt
asgiref==3.8.1
Django==5.1.4
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
